Works, but is slighly slow; runtime ~ 1s
The implementation of the edit distance is O(n * m) with n and m being the lengths of each word.
This is the number of times that we call it:
Each possibly misspelled word (c):
    Each dictionary word (d):
        each letter of the missplled word (n):
            each letter of the dictionary word (m):
                compute something (1)

So, the final time complexity is O(c * d * n * m) This is probably the best O(n) that we will get, but we can make some optimizaitons.
Here are some ideas:
- Tell the edit distance algorthm what the best distance so far is. If at any point the minimum distance is more than that, then this word will not possibly be OK.