added 2 improvements; runtime ~ 450ms
Firstly, I moved the check for duplicates to before anything else in spellcheck(). This means if it is already in the dictionary, it will never do a single edit distance calculation.
Then, I added a parameter for the distance algorithm as somewhat of a quick-stop, to stop a computation early if there is no chance of beating the previous method.
This second optimization halfed the runtime. Much better, but we can keep going.